create multinomial distributions
